requirements.txt

django==2.1.7
djangorestframework==3.9.1
urllib3==1.22
django-oauth-toolkit==1.2.0
django-filter==2.1.0